Linear.Metric

Description

Free metric spaces

Documentation

class Additive f => Metric f where

Free and sparse inner product/metric spaces.

Minimal complete definition

Nothing

Methods

dot :: Num a => f a -> f a -> a

Compute the inner product of two vectors or (equivalently) convert a vector f a into a covector f a -> a.

>>> V2 1 2 `dot` V2 3 4
11

quadrance :: Num a => f a -> a

Compute the squared norm. The name quadrance arises from Norman J. Wildberger's rational trigonometry.

qd :: Num a => f a -> f a -> a

Compute the quadrance of the difference

distance :: Floating a => f a -> f a -> a

Compute the distance between two vectors in a metric space

norm :: Floating a => f a -> a

Compute the norm of a vector in a metric space

signorm :: Floating a => f a -> f a

Convert a non-zero vector to unit vector.

normalize :: (Floating a, Metric f, Epsilon a) => f a -> f a

Normalize a Metric functor to have unit norm. This function does not change the functor if its norm is 0 or 1.

project :: (Metric v, Fractional a) => v a -> v a -> v a

project u v computes the projection of v onto u.