Linear.V

Documentation

data V n a

int :: Int -> TypeQ

This can be used to generate a template haskell splice for a type level version of a given int.

This does not use GHC TypeLits, instead it generates a numeric type by hand similar to the ones used in the "Functional Pearl: Implicit Dimurations" paper by Oleg Kiselyov and Chung-Chieh Shan.

dim :: forall n a. Dim n => V n a -> Int

class Dim n where

Methods

reflectDim :: p n -> Int

reifyDim :: Int -> (forall o. Dim o => Proxy o -> r) -> r

fromVector :: forall n a. Dim n => Vector a -> Maybe (V n a)