Linear.Affine

Description

Operations on affine spaces.

Documentation

class Additive (Diff p) => Affine p where

An affine space is roughly a vector space in which we have forgotten or at least pretend to have forgotten the origin.

a .+^ (b .-. a)  =  b@
(a .+^ u) .+^ v  =  a .+^ (u ^+^ v)@
(a .-. b) ^+^ v  =  (a .+^ v) .-. q@

Minimal complete definition

(.-.), (.+^)

Associated Types

type Diff p :: * -> *

Methods

(.-.) :: Num a => p a -> p a -> Diff p a infixl 6

Get the difference between two points as a vector offset.

(.+^) :: Num a => p a -> Diff p a -> p a infixl 6

Add a vector offset to a point.

(.-^) :: Num a => p a -> Diff p a -> p a infixl 6

Subtract a vector offset from a point.

qdA :: (Affine p, Foldable (Diff p), Num a) => p a -> p a -> a

Compute the quadrance of the difference (the square of the distance)

distanceA :: (Floating a, Foldable (Diff p), Affine p) => p a -> p a -> a

Distance between two points in an affine space

newtype Point f a

A handy wrapper to help distinguish points from vectors at the type level

Constructors

P (f a)

lensP :: Lens' (Point g a) (g a)

_Point :: Iso' (Point f a) (f a)

origin :: (Additive f, Num a) => Point f a

Vector spaces have origins.

relative :: (Additive f, Num a) => Point f a -> Iso' (Point f a) (f a)

An isomorphism between points and vectors, given a reference point.