Linear.Plucker

Description

Plücker coordinates for lines in 3d homogeneous space.

Documentation

data Plucker a

Plücker coordinates for lines in a 3-dimensional space.

Constructors

Plucker !a !a !a !a !a !a

squaredError :: (Eq a, Num a) => Plucker a -> a

Valid Plücker coordinates p will have squaredError p == 0

That said, floating point makes a mockery of this claim, so you may want to use nearZero.

isotropic :: Epsilon a => Plucker a -> Bool

Checks if the line is near-isotropic (isotropic vectors in this quadratic space represent lines in real 3d space).

(><) :: Num a => Plucker a -> Plucker a -> a

This isn't th actual metric because this bilinear form gives rise to an isotropic quadratic space

plucker :: Num a => V4 a -> V4 a -> Plucker a

Given a pair of points represented by homogeneous coordinates generate Plücker coordinates for the line through them, directed from the second towards the first.

plucker3D :: Num a => V3 a -> V3 a -> Plucker a

Given a pair of 3D points, generate Plücker coordinates for the line through them, directed from the second towards the first.

Operations on lines

parallel :: Epsilon a => Plucker a -> Plucker a -> Bool

Checks if two lines are parallel.

intersects :: (Epsilon a, Ord a) => Plucker a -> Plucker a -> Bool

Checks if two lines intersect (or nearly intersect).

data LinePass

Describe how two lines pass each other.

Constructors

Coplanar
The lines are coplanar (parallel or intersecting).

Clockwise
The lines pass each other clockwise (right-handed screw)

Counterclockwise
The lines pass each other counterclockwise (left-handed screw).

passes :: (Epsilon a, Num a, Ord a) => Plucker a -> Plucker a -> LinePass

Check how two lines pass each other. passes l1 l2 describes l2 when looking down l1.

quadranceToOrigin :: Fractional a => Plucker a -> a

The minimum squared distance of a line from the origin.

closestToOrigin :: Fractional a => Plucker a -> V3 a

The point where a line is closest to the origin.

isLine :: Epsilon a => Plucker a -> Bool

Not all 6-dimensional points correspond to a line in 3D. This predicate tests that a Plücker coordinate lies on the Grassmann manifold, and does indeed represent a 3D line.

data Coincides a where

When lines are represented as Plücker coordinates, we have the ability to check for both directed and undirected equality. Undirected equality between Lines (or a Line and a Ray) checks that the two lines coincide in 3D space. Directed equality, between two Rays, checks that two lines coincide in 3D, and have the same direction. To accomodate these two notions of equality, we use an Eq instance on the Coincides data type.

For example, to check the directed equality between two lines, p1 and p2, we write, Ray p1 == Ray p2.

Constructors

Line :: (Epsilon a, Fractional a) => Plucker a -> Coincides a

Ray :: (Epsilon a, Fractional a, Ord a) => Plucker a -> Coincides a

Basis elements

p01 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p02 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p03 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p10 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a

p12 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p13 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a

p20 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a

p21 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a

p23 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p30 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a

p31 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p32 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a