Linear.Plucker

Description

Plücker coordinates for lines in 3d homogeneous space.

Documentation

data Plucker a

Plücker coordinates for lines in a 3-dimensional space.

Constructors

Plucker a a a a a a

squaredError :: (Eq a, Num a) => Plucker a -> a

Valid Plücker coordinates p will have squaredError p == 0

That said, floating point makes a mockery of this claim, so you may want to use nearZero.

isotropic :: Epsilon a => Plucker a -> Bool

Checks if the line is near-isotropic (isotropic vectors in this quadratic space represent lines in real 3d space)

(><) :: Num a => Plucker a -> Plucker a -> a

This isn't th actual metric because this bilinear form gives rise to an isotropic quadratic space

plucker :: Num a => V4 a -> V4 a -> Plucker a

Given a pair of points represented by homogeneous coordinates generate Plücker coordinates for the line through them.

intersects :: Epsilon a => Plucker a -> Plucker a -> Bool

Checks if the two vectors intersect (or nearly intersect)

Basis elements

p01, p12, p31, p23, p03, p02 :: Functor f => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form a basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p01 :: Lens' (Plucker a) a
 p02 :: Lens' (Plucker a) a
 p03 :: Lens' (Plucker a) a
 p23 :: Lens' (Plucker a) a
 p31 :: Lens' (Plucker a) a
 p12 :: Lens' (Plucker a) a

p10, p21, p13, p32, p30, p20 :: (Functor f, Num a) => (a -> f a) -> Plucker a -> f (Plucker a)

These elements form an alternate basis for the Plücker space, or the Grassmanian manifold Gr(2,V4).

 p10 :: Num a => Lens' (Plucker a) a
 p20 :: Num a => Lens' (Plucker a) a
 p30 :: Num a => Lens' (Plucker a) a
 p32 :: Num a => Lens' (Plucker a) a
 p13 :: Num a => Lens' (Plucker a) a
 p21 :: Num a => Lens' (Plucker a) a